+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 [Semantic Versioning](https://semver.org/)
 
-## [2.0.0] - 2021-06-21
-### Code refactor and deprecations
+## [2.0.0] - 2021-07-11
+### Code refactor, deprecations and bug fixes
 - Inheritance to composition for collector classes
 - async/await promise style
 - Recursive waits migrated to rxjx Observables
 - PFV unique building method
-- RcsbFvWebApp.setBoardConfig has been removed. Custom board config must be defined in RcsbFvAdditionalConfig. Some RcsbFvWebApp methods do not support additional configs yet
+- RcsbFvWebApp.setBoardConfig  and RcsbFvWebApp.getRcsbFv have been removed. 
+  - Custom board config must be defined in RcsbFvAdditionalConfig (Some RcsbFvWebApp methods do not support additional configs yet)
+  - Builder methods return RcsbFvModulePublicInterface
+- Context Manager rcsbFv build new board config param 
+- Select button default value title bug fixed
 
 ## [1.4.8] - 2021-06-15
 ### Optimization